implemented stop process func

+++ beast/src/main/java/edu/pse/beast/api/testrunner/propertycheck/processes/process_handler/CBMCProcessHandlerLinux.java
@@ -41,8 +41,9 @@ public class CBMCProcessHandlerLinux implements CBMCProcessHandler {
 
 	@Override
 	public void endProcess(Process p) {
+		if (!p.isAlive())
+			return;
 		p.destroyForcibly();
 	}
-
 
 }
